Se soluciono tema de dinero como float no como int

+++ app/src/main/java/com/example/nicolas/clientefinalandroid2/Activities/Tabs/ActivityCargaJugada.java
@@ -88,7 +88,7 @@ public class ActivityCargaJugada extends ActionBarActivity implements View.OnCli
         if(!jugadaProvisoria.estoyVacia())
         {
             tv1.setText(jugadaProvisoria.getNumero());
-            posicionSpinner =jugadaProvisoria.getDineroApostado();
+            posicionSpinner =(int)jugadaProvisoria.getDineroApostado();
             dameSuenio();
 
         }
@@ -270,14 +270,16 @@ public class ActivityCargaJugada extends ActionBarActivity implements View.OnCli
 
 
     @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
+    public boolean onCreateOptionsMenu(Menu menu)
+    {
         // Inflate the menu; this adds items to the action bar if it is present.
         getMenuInflater().inflate(R.menu.menu_tab1, menu);
         return true;
     }
 
     @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
+    public boolean onOptionsItemSelected(MenuItem item)
+    {
         // Handle action bar item clicks here. The action bar will
         // automatically handle clicks on the Home/Up button, so long
         // as you specify a parent activity in AndroidManifest.xml.
